+++ core/v2/aggregation.go
