+++ core/v2/aggregation_test.go
@@ -7,7 +7,7 @@ import (
 	"testing"
 
 	"github.com/Layr-Labs/eigenda/chainio/mock"
-	corev2 "github.com/Layr-Labs/eigenda/corev2"
+	corev2 "github.com/Layr-Labs/eigenda/core/v2"
 	"github.com/Layr-Labs/eigenda/crypto/ecc/bn254"
 	"github.com/stretchr/testify/assert"
 )
